# More generous detection of little endian systems