Fixes https://github.com/RoaringBitmap/roaring/issues/284