@lemire 2020-10-29
I think we are good now. Merging.