@jacksonrnewhouse 2020-10-29
For posterity, here are the before and after benchmarks on an AWS Neoverse machine. Anyone running on an affected platfrom should see massive improvements.

```
benchstat --delta-test none -sort -delta  old.txt new.txt
name                                                                  old time/op    new time/op    delta
SerializationDense-16                                                   29.6µs ± 0%     0.9µs ± 0%  -96.92%
UnserializeFromBuffer/FromBuffer-6500-16                                6.73µs ± 0%    0.22µs ± 0%  -96.73%
UnserializeFromBuffer/FromBuffer-65000-16                               6.71µs ± 0%    0.22µs ± 0%  -96.72%
SerializationMid-16                                                      210µs ± 0%      14µs ± 0%  -93.24%
UnserializeFromBuffer/FromBuffer-650-16                                 2.16µs ± 0%    0.22µs ± 0%  -89.83%
UnserializeFromBuffer/FromBuffer-650000-16                              8.15µs ± 0%    1.10µs ± 0%  -86.50%
UnmarshalBinary-16                                                       321µs ± 0%      69µs ± 0%  -78.59%
MarshalBinary-16                                                         371µs ± 0%      79µs ± 0%  -78.56%
UnserializeReadFrom/ReadFrom-65000-16                                   12.5µs ± 0%     3.2µs ± 0%  -74.79%
UnserializeReadFrom/ReadFrom-6500-16                                    12.5µs ± 0%     3.2µs ± 0%  -74.75%
SerializationSparse-16                                                   330µs ± 0%      89µs ± 0%  -73.11%
UnserializeReadFrom/ReadFrom-650-16                                     3.45µs ± 0%    0.96µs ± 0%  -72.24%
UnserializeReadFrom/ReadFrom-650000-16                                  15.1µs ± 0%     5.0µs ± 0%  -66.84%
UnserializeFromBuffer/FromBuffer-6500000-16                             21.2µs ± 0%     8.9µs ± 0%  -58.07%
UnserializeReadFrom/ReadFrom-6500000-16                                 36.4µs ± 0%    20.3µs ± 0%  -44.03%
AndAny/small-filters_or-first-16                                        1.35ms ± 0%    0.93ms ± 0%  -31.54%
ArrayUnionThreshold/mostly-overlap-16                                   22.0µs ± 0%    15.3µs ± 0%  -30.55%
ArrayUnionThreshold/little-overlap-16                                   24.6µs ± 0%    18.0µs ± 0%  -26.66%
XorLopsided-16                                                           326µs ± 0%     243µs ± 0%  -25.41%
ArrayUnionThreshold/no-overlap-16                                       41.7µs ± 0%    33.1µs ± 0%  -20.76%
MarshalMsgaddHelper16-16                                                 107ns ± 0%      86ns ± 0%  -19.63%
MarshalMsgroaringArray-16                                                106ns ± 0%      85ns ± 0%  -19.62%
MarshalMsgbitmapContainer-16                                            76.5ns ± 0%    61.6ns ± 0%  -19.48%
MarshalMsgrunIterator16-16                                              84.9ns ± 0%    68.8ns ± 0%  -18.96%
FromBitmap16-16                                                         18.5µs ± 0%    15.2µs ± 0%  -17.55%
MarshalMsgbitmapContainerShortIterator-16                               58.0ns ± 0%    48.1ns ± 0%  -17.07%
MarshalMsgrunContainer16-16                                             66.9ns ± 0%    55.8ns ± 0%  -16.59%
MarshalMsginterval16-16                                                 64.5ns ± 0%    54.0ns ± 0%  -16.28%
ShortIteratorAdvanceRuntime/init_size_1_shortIterator_advance-16        73.6ns ± 0%    63.0ns ± 0%  -14.40%
IntersectionLargeParallel-16                                            2.90ms ± 0%    2.51ms ± 0%  -13.65%
ShortIteratorNextRuntime/init_size_1_shortIterator_next-16              76.9ns ± 0%    66.5ns ± 0%  -13.52%
Xor-16                                                                  1.61ms ± 0%    1.40ms ± 0%  -13.12%
ShortIteratorAdvanceRuntime/init_size_65535_shortIterator_advance-16    83.6ns ± 0%    72.9ns ± 0%  -12.80%
ShortIteratorAdvanceRuntime/init_size_650_shortIterator_advance-16      83.3ns ± 0%    72.7ns ± 0%  -12.73%
ShortIteratorAdvanceRuntime/init_size_6500_shortIterator_advance-16     83.4ns ± 0%    72.8ns ± 0%  -12.71%
AntagonisticArrayUnionsGrowth-16                                        59.0µs ± 0%    51.7µs ± 0%  -12.42%
SparseAdvanceRoaring/advance_from_65-16                                  163ns ± 0%     143ns ± 0%  -12.27%
SparseAdvanceRoaring/advance_from_650-16                                 164ns ± 0%     144ns ± 0%  -12.20%
SparseAdvanceRoaring/advance_from_1-16                                   162ns ± 0%     143ns ± 0%  -11.73%
SparseAdvanceRoaring/advance_from_2-16                                   162ns ± 0%     143ns ± 0%  -11.73%
UnionInPlaceCopyOnWrite/10-4-3-3-16                                     9.66µs ± 0%    8.56µs ± 0%  -11.42%
UnionInPlaceCopyOnWrite/7-4-3-3-16                                      4.44µs ± 0%    3.97µs ± 0%  -10.65%
ShortIteratorAdvanceBitmap/init_size_1_shortIterator_advance-16         56.9ns ± 0%    52.1ns ± 0%   -8.44%
Nexts/next__1.562500%-16                                                8.26ms ± 0%    7.64ms ± 0%   -7.43%
ShortIteratorAdvanceBitmap/init_size_65535_shortIterator_advance-16     65.3ns ± 0%    60.5ns ± 0%   -7.35%
Nexts/nextmany__0.012352%-16                                            6.96ms ± 0%    6.45ms ± 0%   -7.34%
ShortIteratorAdvanceBitmap/init_size_650_shortIterator_advance-16       65.5ns ± 0%    60.7ns ± 0%   -7.33%
AndAny/equal_and-first-16                                               11.8ms ± 0%    10.9ms ± 0%   -7.13%
ShortIteratorAdvanceBitmap/init_size_6500_shortIterator_advance-16      65.1ns ± 0%    60.5ns ± 0%   -7.07%
Nexts/next__6.250000%-16                                                8.30ms ± 0%    7.72ms ± 0%   -7.01%
AndAny/small-filters_and-first-16                                        131µs ± 0%     122µs ± 0%   -6.48%
MarshalMsgcontainerSerz-16                                              42.4ns ± 0%    39.7ns ± 0%   -6.37%
BitmapReuseWithoutClear-16                                              4.42ms ± 0%    4.14ms ± 0%   -6.36%
AndAny/equal_or-first-16                                                29.9ms ± 0%    28.0ms ± 0%   -6.19%
MarshalMsguint16Slice-16                                                30.7ns ± 0%    28.9ns ± 0%   -5.86%
SparseContains-16                                                        109µs ± 0%     103µs ± 0%   -5.72%
Nexts/next__3.125000%-16                                                8.22ms ± 0%    7.76ms ± 0%   -5.50%
Nexts/nextmany__0.097656%-16                                            2.45ms ± 0%    2.32ms ± 0%   -5.37%
AndAny/small-base_or-first-16                                           47.3ms ± 0%    45.2ms ± 0%   -4.46%
BitmapReuseWithClear-16                                                 4.14ms ± 0%    3.95ms ± 0%   -4.43%
IntersectionLargeRoaring-16                                             5.63ms ± 0%    5.40ms ± 0%   -4.17%
Nexts/next__25.000000%-16                                               9.93ms ± 0%    9.63ms ± 0%   -3.10%
UnionInPlaceCopyOnWrite/10-7-3-3-16                                     33.7µs ± 0%    32.8µs ± 0%   -2.76%
AndAny/small-filters_AndAny-16                                           182µs ± 0%     177µs ± 0%   -2.59%
Nexts/next__0.012352%-16                                                12.6ms ± 0%    12.4ms ± 0%   -1.93%
Nexts/next__50.000000%-16                                               9.61ms ± 0%    9.43ms ± 0%   -1.92%
Unmarshalinterval16-16                                                  65.3ns ± 0%    64.1ns ± 0%   -1.84%
Unmarshaluint16Slice-16                                                 11.1ns ± 0%    10.9ns ± 0%   -1.80%
ShortIteratorAdvanceArray/init_size_6500_shortIterator_advance-16        116ns ± 0%     114ns ± 0%   -1.72%
ShortIteratorNextArray/init_size_65535_shortIterator_next-16             407µs ± 0%     400µs ± 0%   -1.71%
Nexts/nextmany__0.390625%-16                                            1.89ms ± 0%    1.86ms ± 0%   -1.58%
AppendMsgrunIterator16-16                                               25.9ns ± 0%    25.5ns ± 0%   -1.54%
EncoderunIterator16-16                                                  72.3ns ± 0%    71.2ns ± 0%   -1.52%
UnionInPlaceCopyOnWrite/4-7-3-3-16                                      7.56µs ± 0%    7.45µs ± 0%   -1.46%
UnionInPlaceCopyOnWrite/7-7-3-3-16                                      27.8µs ± 0%    27.5µs ± 0%   -1.40%
EncodecontainerSerz-16                                                  45.9ns ± 0%    45.4ns ± 0%   -1.09%
EqualsSparse-16                                                         91.6µs ± 0%    90.7µs ± 0%   -0.98%
IterateRoaring/iterator-16                                               973µs ± 0%     963µs ± 0%   -0.98%
CountRoaring-16                                                         66.9ns ± 0%    66.3ns ± 0%   -0.90%
DecoderoaringArray-16                                                    224ns ± 0%     222ns ± 0%   -0.89%
DecoderunContainer16-16                                                  118ns ± 0%     117ns ± 0%   -0.85%
ShortIteratorAdvanceArray/init_size_65535_shortIterator_advance-16       125ns ± 0%     124ns ± 0%   -0.80%
SparseIterateBitset-16                                                  2.71ms ± 0%    2.69ms ± 0%   -0.78%
Nexts/nextmany__1.562500%-16                                            1.76ms ± 0%    1.75ms ± 0%   -0.77%
NewBitmap-16                                                            67.3ns ± 0%    66.8ns ± 0%   -0.74%
UnionInPlaceCopyOnWrite/4-10-3-3-16                                     37.5µs ± 0%    37.2µs ± 0%   -0.74%
UnmarshalcontainerSerz-16                                               81.9ns ± 0%    81.3ns ± 0%   -0.73%
UnionBitset-16                                                          2.91ms ± 0%    2.89ms ± 0%   -0.64%
UnionInPlaceCopyOnWrite/10-10-3-10-16                                   3.62ms ± 0%    3.60ms ± 0%   -0.64%
Nexts/nextmany__3.125000%-16                                            1.74ms ± 0%    1.73ms ± 0%   -0.60%
Nexts/nextmany__6.250000%-16                                            1.72ms ± 0%    1.71ms ± 0%   -0.54%
Encodeinterval16-16                                                     49.9ns ± 0%    49.7ns ± 0%   -0.40%
Nexts/nextmany__100.000000%-16                                          1.53ms ± 0%    1.52ms ± 0%   -0.39%
IntersectionBitset-16                                                   7.82µs ± 0%    7.79µs ± 0%   -0.37%
UnmarshalbitmapContainer-16                                             54.6ns ± 0%    54.4ns ± 0%   -0.37%
ShortIteratorNextArray/init_size_6500_shortIterator_next-16             39.6µs ± 0%    39.5µs ± 0%   -0.35%
UnionRoaring-16                                                          291µs ± 0%     290µs ± 0%   -0.35%
Nexts/nextmany__25.000000%-16                                           2.19ms ± 0%    2.18ms ± 0%   -0.28%
IterateRoaring/iterator-compressed-16                                    972µs ± 0%     970µs ± 0%   -0.23%
AppendMsgaddHelper16-16                                                 42.9ns ± 0%    42.8ns ± 0%   -0.23%
ShortIteratorNextBitmap/init_size_1_shortIterator_next-16               1.30µs ± 0%    1.30µs ± 0%   -0.23%
NextsRLE/nextmany-16                                                    2.93ms ± 0%    2.92ms ± 0%   -0.22%
UnmarshalbitmapContainerShortIterator-16                                49.2ns ± 0%    49.1ns ± 0%   -0.20%
Nexts/next__12.500000%-16                                               10.3ms ± 0%    10.3ms ± 0%   -0.19%
UnmarshalrunContainer16-16                                              54.6ns ± 0%    54.5ns ± 0%   -0.18%
Nexts/nextmany__12.500000%-16                                           2.26ms ± 0%    2.25ms ± 0%   -0.18%
ShortIteratorAdvanceArray/init_size_1_shortIterator_advance-16          58.3ns ± 0%    58.2ns ± 0%   -0.17%
SetRoaring-16                                                           59.1ns ± 0%    59.0ns ± 0%   -0.17%
ShortIteratorNextBitmap/init_size_6500_shortIterator_next-16            75.1µs ± 0%    75.0µs ± 0%   -0.16%
SparseAdvanceSequentially/advance_from_65-16                             785µs ± 0%     784µs ± 0%   -0.14%
Nexts/nextmany__50.000000%-16                                           2.27ms ± 0%    2.27ms ± 0%   -0.14%
ShortIteratorNextRuntime/init_size_650_shortIterator_next-16            6.43µs ± 0%    6.42µs ± 0%   -0.12%
EmptyArray-16                                                           8.20ns ± 0%    8.19ns ± 0%   -0.12%
UnionInPlaceCopyOnWrite/7-10-3-3-16                                     57.2µs ± 0%    57.2µs ± 0%   -0.11%
IntersectionRoaring-16                                                  1.11µs ± 0%    1.11µs ± 0%   -0.09%
SparseAdvanceSequentially/advance_from_2-16                              242µs ± 0%     241µs ± 0%   -0.08%
ShortIteratorNextBitmap/init_size_65535_shortIterator_next-16            737µs ± 0%     737µs ± 0%   -0.07%
ShortIteratorNextArray/init_size_650_shortIterator_next-16              4.02µs ± 0%    4.01µs ± 0%   -0.05%
Ors/fast-16                                                             5.32ms ± 0%    5.32ms ± 0%   -0.05%
IterateRoaring/iterate-16                                                590µs ± 0%     590µs ± 0%   -0.04%
CountBitset-16                                                          17.8µs ± 0%    17.8µs ± 0%   -0.01%
MarshalMsgarrayContainer-16                                             40.0ns ± 0%    40.0ns ± 0%    0.00%
AppendMsgarrayContainer-16                                              14.2ns ± 0%    14.2ns ± 0%    0.00%
UnmarshalarrayContainer-16                                              32.6ns ± 0%    32.6ns ± 0%    0.00%
EncodearrayContainer-16                                                 28.7ns ± 0%    28.7ns ± 0%    0.00%
DecodearrayContainer-16                                                 69.9ns ± 0%    69.9ns ± 0%    0.00%
ShortIteratorAdvanceArray/init_size_650_shortIterator_advance-16         104ns ± 0%     104ns ± 0%    0.00%
MemoryUsage-16                                                          0.00ns ± 0%    0.00ns ± 0%    0.00%
SetBitset-16                                                            22.2ns ± 0%    22.2ns ± 0%    0.00%
GetTestRoaring-16                                                        121ns ± 0%     121ns ± 0%    0.00%
AppendMsgbitmapContainer-16                                             24.8ns ± 0%    24.8ns ± 0%    0.00%
EncodebitmapContainer-16                                                53.6ns ± 0%    53.6ns ± 0%    0.00%
DecodebitmapContainer-16                                                 119ns ± 0%     119ns ± 0%    0.00%
100CountTrailingZeros-16                                                60.6ns ± 0%    60.6ns ± 0%    0.00%
Popcount-16                                                             74.7ns ± 0%    74.7ns ± 0%    0.00%
DecodecontainerSerz-16                                                   140ns ± 0%     140ns ± 0%    0.00%
EncodeaddHelper16-16                                                     114ns ± 0%     114ns ± 0%    0.00%
DecodeaddHelper16-16                                                     283ns ± 0%     283ns ± 0%    0.00%
AppendMsginterval16-16                                                  19.6ns ± 0%    19.6ns ± 0%    0.00%
Decodeinterval16-16                                                      125ns ± 0%     125ns ± 0%    0.00%
UnmarshalrunIterator16-16                                               74.9ns ± 0%    74.9ns ± 0%    0.00%
DecoderunIterator16-16                                                   176ns ± 0%     176ns ± 0%    0.00%
AppendMsguint16Slice-16                                                 10.1ns ± 0%    10.1ns ± 0%    0.00%
Encodeuint16Slice-16                                                    13.8ns ± 0%    13.8ns ± 0%    0.00%
NextsRLE/next-16                                                        17.7ms ± 0%    17.7ms ± 0%   +0.02%
EqualsClone-16                                                          88.2µs ± 0%    88.2µs ± 0%   +0.02%
Ors/two-by-two-16                                                        741µs ± 0%     741µs ± 0%   +0.03%
IterateRoaring/iterate-compressed-16                                     590µs ± 0%     590µs ± 0%   +0.07%
SparseAdvanceSequentially/advance_from_1-16                              121µs ± 0%     121µs ± 0%   +0.09%
ShortIteratorNextArray/init_size_1_shortIterator_next-16                58.3ns ± 0%    58.4ns ± 0%   +0.17%
EncoderunContainer16-16                                                 46.0ns ± 0%    46.1ns ± 0%   +0.22%
AppendMsgroaringArray-16                                                40.8ns ± 0%    40.9ns ± 0%   +0.25%
SparseAdvanceSequentially/advance_from_650-16                            769µs ± 0%     771µs ± 0%   +0.28%
UnmarshalroaringArray-16                                                96.4ns ± 0%    96.7ns ± 0%   +0.31%
ShortIteratorNextBitmap/init_size_650_shortIterator_next-16             8.63µs ± 0%    8.66µs ± 0%   +0.31%
IterateBitset-16                                                         406µs ± 0%     407µs ± 0%   +0.34%
Nexts/next__100.000000%-16                                              8.83ms ± 0%    8.86ms ± 0%   +0.35%
ShortIteratorNextRuntime/init_size_65535_shortIterator_next-16           640µs ± 0%     642µs ± 0%   +0.36%
Ors/next/add-16                                                         1.00ms ± 0%    1.01ms ± 0%   +0.43%
Decodeuint16Slice-16                                                    21.6ns ± 0%    21.7ns ± 0%   +0.46%
AppendMsgrunContainer16-16                                              19.8ns ± 0%    19.9ns ± 0%   +0.51%
AppendMsgcontainerSerz-16                                               19.0ns ± 0%    19.1ns ± 0%   +0.53%
UnionInPlaceCopyOnWrite/10-4-3-10-16                                    61.5µs ± 0%    61.8µs ± 0%   +0.56%
UnionInPlaceCopyOnWrite/10-7-3-10-16                                     451µs ± 0%     453µs ± 0%   +0.62%
AppendMsgbitmapContainerShortIterator-16                                13.4ns ± 0%    13.5ns ± 0%   +0.75%
GetTestBitSet-16                                                        13.2ns ± 0%    13.3ns ± 0%   +0.76%
DecodebitmapContainerShortIterator-16                                    122ns ± 0%     123ns ± 0%   +0.82%
AndAny/small-base_and-first-16                                           160µs ± 0%     162µs ± 0%   +0.83%
ShortIteratorNextRuntime/init_size_6500_shortIterator_next-16           63.3µs ± 0%    63.8µs ± 0%   +0.84%
UnionInPlaceCopyOnWrite/7-10-3-10-16                                     481µs ± 0%     485µs ± 0%   +0.85%
UnionInPlaceCopyOnWrite/4-10-3-10-16                                    89.1µs ± 0%    89.9µs ± 0%   +0.88%
EncodebitmapContainerShortIterator-16                                   43.1ns ± 0%    43.5ns ± 0%   +0.93%
AndAny/equal_AndAny-16                                                  16.0ms ± 0%    16.2ms ± 0%   +1.14%
Nexts/next__0.390625%-16                                                8.31ms ± 0%    8.41ms ± 0%   +1.22%
UnionInPlaceCopyOnWrite/7-4-3-10-16                                     55.6µs ± 0%    56.3µs ± 0%   +1.24%
UnionInPlaceCopyOnWrite/10-10-10-10-16                                  13.4ms ± 0%    13.5ms ± 0%   +1.25%
UnionInPlaceCopyOnWrite/10-10-3-3-16                                     229µs ± 0%     232µs ± 0%   +1.33%
SparseAdvanceOnHugeData-16                                              1.10µs ± 0%    1.11µs ± 0%   +1.55%
SparseIterateRoaring-16                                                 1.08ms ± 0%    1.10ms ± 0%   +1.58%
UnionInPlaceCopyOnWrite/4-10-10-10-16                                    240µs ± 0%     244µs ± 0%   +1.65%
…